Double-hashed Denylist mention
While doing some blogpost work, I've come back to this and noticed that there is a mention of double-hashed denylists. I remember that we discussed this back then and learned that denylists are not double-hashed (anymore). Since then, I was left with the impression that the text of this blogpost has changed accordingly, but it turns out that it hasn't, hence, this small PR is doing a small rephrase.

@daviddias @momack2 can you please see if you agree and merge? Or if I got it wrong, please ignore.