Switch to ISO 8601 dates
Currently the blog is using dates in the form of "year.month.day".
This is confusing for folks like me, who come from an area where
dates are also written with a dot as separator, but a different
order: "day.month.year".

Hence it makes sense to display dates in the format of the ISO 8601
standard: year-month-day (with leading zeros).